# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Study-ing API
  version: 0.2.0
  description: |
    Study-ing 프론트엔드와 Spring Boot 백엔드가 공유하는 API 계약입니다.

    구현 원칙:
    - GitLab access token은 브라우저에 반환하지 않습니다.
    - 인증된 API는 HttpOnly 세션 쿠키를 사용합니다.
    - Workspace ID로 연결 프로젝트를 결정하며 요청에서 임의 프로젝트를 받지 않습니다.
    - 날짜는 `YYYY-MM-DD`, 시각은 timezone을 포함한 ISO 8601 문자열입니다.
    - 수정 요청은 `expectedRevision` 또는 `lastCommitId`로 낙관적 잠금을 적용합니다.
  contact:
    name: Study-ing Team
servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
...
paths:
...
  /gitlab/projects:
    get:
      tags: [GitLab Projects]
      summary: 현재 사용자가 접근 가능한 GitLab 프로젝트 목록
      operationId: listGitLabProjects
      x-owner: member-1
      x-implementation-status: implemented
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: 프로젝트 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GitLabProject"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "502":
          $ref: "#/components/responses/GitLabError"
  /gitlab/projects/{projectId}/connection-check:
    get:
      tags: [GitLab Projects]
      summary: 프로젝트 연결 가능 여부 확인
      operationId: checkGitLabProjectConnection
      x-owner: member-1
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OAuth 사용자의 프로젝트 조회 성공 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitLabConnectionSpike"
        "403":
          $ref: "#/components/responses/GitLabAccessDenied"
        "502":
          $ref: "#/components/responses/GitLabError"
  /gitlab/projects/{projectId}/repository/file:
    get:
      tags: [GitLab Projects]
      summary: OAuth 사용자 권한으로 프로젝트 파일 조회
      operationId: getOAuthGitLabProjectFile
      x-owner: member-1
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: path
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: ref
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Base64를 서버에서 디코딩한 텍스트 파일
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitLabFileContent"
        "400":
          $ref: "#/components/responses/InvalidPath"
        "403":
          $ref: "#/components/responses/GitLabAccessDenied"
        "502":
          $ref: "#/components/responses/GitLabError"
  /gitlab/projects/{projectId}/import-analysis:
    get:
      tags: [GitLab Projects]
      summary: 기존 저장소를 변경하지 않고 가져오기 가능 여부 분석
      operationId: analyzeGitLabRepository
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 저장소 분류, 호환 파일 수, 전용 경로 충돌과 tree fingerprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryImportAnalysis"
        "403":
          $ref: "#/components/responses/GitLabAccessDenied"
        "413":
          $ref: "#/components/responses/InvalidRequest"
...
  /workspaces/{workspaceId}/sessions/{date}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/SessionDate"
    get:
      tags: [Sessions]
      summary: 특정 날짜 일정
      operationId: getSession
      x-owner: member-2
      x-implementation-status: implemented
      responses:
        "200":
          description: 일정 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Sessions]
      summary: revision과 lastCommitId를 확인하고 일정 수정
      operationId: updateSession
      x-owner: member-2
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: 수정된 일정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Sessions]
      summary: session.yml을 삭제하지 않고 cancelled 처리
      operationId: cancelSession
      x-owner: member-2
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedRevision, lastCommitId, changeReason]
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                lastCommitId:
                  type: string
                changeReason:
                  type: string
                  minLength: 1
                  maxLength: 300
      responses:
        "200":
          description: cancelled 상태의 일정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /workspaces/{workspaceId}/repository/tree:
    get:
      tags: [Repository]
      summary: 연결 프로젝트의 repository tree
      operationId: getRepositoryTree
      x-owner: member-2
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: path
          in: query
          schema:
            type: string
            default: ""
        - name: ref
          in: query
          schema:
            type: string
          description: 생략하면 Workspace 기본 브랜치
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: 저장소 항목
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryTreeResponse"
        "403":
          $ref: "#/components/responses/InvalidPath"
  /workspaces/{workspaceId}/repository/file:
    get:
      tags: [Repository]
      summary: 허용된 텍스트 파일과 커밋 정보
      operationId: getRepositoryFile
      x-owner: member-2
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 파일 미리보기
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryFileResponse"
        "403":
          $ref: "#/components/responses/InvalidPath"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SessionDate:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    InvalidRequest:
      description: 요청 형식 또는 비즈니스 입력값 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: INVALID_REQUEST
            message: 요청값을 확인해 주세요.
            details:
              field: title
            requestId: req_01J
            timestamp: "2026-07-25T12:00:00Z"
    InvalidPath:
      description: 허용되지 않은 저장소 경로
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: FILE_PATH_NOT_ALLOWED
            message: 허용되지 않은 저장소 경로입니다.
            timestamp: "2026-07-25T12:00:00Z"
    AuthRequired:
      description: 로그인 또는 유효한 세션 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: AUTH_REQUIRED
            message: 로그인이 필요합니다.
            timestamp: "2026-07-25T12:00:00Z"
...
    GitLabAccessDenied:
      description: GitLab 프로젝트 권한 부족
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: GITLAB_PROJECT_ACCESS_DENIED
            message: GitLab 프로젝트에 접근할 수 없습니다.
            timestamp: "2026-07-25T12:00:00Z"
    NotFound:
      description: Workspace, Session, Item 또는 파일을 찾지 못함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: RESOURCE_NOT_FOUND
            message: 요청한 리소스를 찾지 못했습니다.
            timestamp: "2026-07-25T12:00:00Z"
    Conflict:
      description: revision, lastCommitId 또는 중복 리소스 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: SESSION_REVISION_CONFLICT
            message: 다른 사용자가 먼저 수정했습니다. 최신 내용을 확인해 주세요.
            details:
              expectedRevision: 3
              actualRevision: 4
            timestamp: "2026-07-25T12:00:00Z"
...
    GitLabError:
      description: GitLab 인증·권한·네트워크·요청 제한 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: GITLAB_API_ERROR
            message: GitLab 요청을 처리하지 못했습니다.
            requestId: req_01J
            timestamp: "2026-07-25T12:00:00Z"
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: 프론트 분기와 테스트에서 사용하는 안정적인 오류 코드
        message:
          type: string
          description: 사용자에게 표시 가능한 한국어 메시지
        details:
          type: object
          additionalProperties: true
          description: 필드 오류나 충돌 최신값
        requestId:
          type: string
          description: 로그 추적 ID
        timestamp:
          type: string
          format: date-time
...
    PageMeta:
      type: object
      required: [page, perPage, total]
      properties:
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
...
    RepositoryImportAnalysis:
      type: object
      required: [projectId, classification, repositoryBasePath, repositorySchemaVersion, treeFingerprint, totalFiles, compatibleSessions, compatibleSubmissions, ignoredFiles, issues]
      properties:
        projectId: { type: integer, format: int64 }
        projectPath: { type: string }
        defaultBranch: { type: string }
        classification:
          type: string
          enum: [EMPTY, COMPATIBLE, LEGACY, PARTIALLY_COMPATIBLE, CONFLICTED]
        repositoryBasePath: { type: string }
        repositorySchemaVersion: { type: integer, enum: [1, 2] }
        treeFingerprint: { type: string, minLength: 64, maxLength: 64 }
        totalFiles: { type: integer, minimum: 0 }
        compatibleSessions: { type: integer, minimum: 0 }
        compatibleSubmissions: { type: integer, minimum: 0 }
        ignoredFiles: { type: integer, minimum: 0 }
        issues:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              code: { type: string }
              message: { type: string }
    GitLabProject:
      type: object
      required: [id, name, pathWithNamespace, visibility]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        pathWithNamespace:
          type: string
        defaultBranch:
          type: [string, "null"]
        visibility:
          type: string
        webUrl:
          type: string
          format: uri
        effectiveAccessLevel:
          type: integer
          description: GitLab access level, Developer는 30 이상
...
    SessionType:
      type: string
      enum: [algorithm, english, cs, free]
    SessionStatus:
      type: string
      enum: [active, cancelled]
    SubmitType:
      type: string
      enum: [link, text, code, mixed]
    StudyItem:
      type: object
      required: [id, order, title, type, submitType, required, status]
      properties:
        id:
          type: string
          pattern: ^item-[a-f0-9]{6,16}$
        order:
          type: integer
          minimum: 1
        title:
          type: string
        type:
          $ref: "#/components/schemas/SessionType"
        source:
          type: [string, "null"]
        url:
          type: [string, "null"]
          format: uri
        submitType:
          $ref: "#/components/schemas/SubmitType"
        required:
          type: boolean
        status:
          type: string
          enum: [active, cancelled]
    StudySession:
      type: object
      required:
        [date, folder, revision, type, title, description, status, deadline, items, lastCommitId]
      properties:
        date:
          type: string
          format: date
        folder:
          type: string
          pattern: ^\d{6}$
        revision:
          type: integer
          minimum: 1
        type:
          $ref: "#/components/schemas/SessionType"
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        deadline:
          type: string
          format: date-time
        secondaryDeadline:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudyItem"
        lastCommitId:
          type: string
    SessionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/StudySession"
    SessionItemDraft:
      type: object
      required: [title, type, submitType, required]
      properties:
        id:
          type: string
          description: 수정 시 기존 항목 유지에만 전달
        title:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: "#/components/schemas/SessionType"
        source:
          type: string
          maxLength: 80
        url:
          type: string
          format: uri
        submitType:
          $ref: "#/components/schemas/SubmitType"
        required:
          type: boolean
    CreateSessionRequest:
      type: object
      required: [date, type, title, description, deadline, items]
      properties:
        date:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/SessionType"
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        deadline:
          type: string
          format: date-time
        secondaryDeadline:
          type: [string, "null"]
          format: date-time
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SessionItemDraft"
    UpdateSessionRequest:
      allOf:
        - $ref: "#/components/schemas/CreateSessionRequest"
        - type: object
          required: [expectedRevision, lastCommitId, changeReason]
          properties:
            expectedRevision:
              type: integer
              minimum: 1
            lastCommitId:
              type: string
            changeReason:
              type: string
              minLength: 1
              maxLength: 300
    RepositoryTreeItem:
      type: object
      required: [id, name, type, path, mode]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [blob, tree]
        path:
          type: string
        mode:
          type: string
    RepositoryTreeResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryTreeItem"
        meta:
          $ref: "#/components/schemas/PageMeta"
    RepositoryFileResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [path, type, rawContent, lastCommitId, webUrl]
          properties:
            path:
              type: string
            type:
              type: string
              enum: [yaml, markdown, text, code]
            rawContent:
              type: string
            parsedContent:
              type: [object, "null"]
              additionalProperties: true
            lastCommitId:
              type: string
            webUrl:
              type: string
              format: uri
...
    GitLabConnectionSpike:
      type: object
      required: [configured, status, message, checkedAt, repositoryTree]
      properties:
        configured:
          type: boolean
        status:
          type: string
          enum: [CONNECTED, NOT_CONFIGURED]
        message:
          type: string
        checkedAt:
          type: string
          format: date-time
        user:
          type: [object, "null"]
          additionalProperties: true
        project:
          type: [object, "null"]
          additionalProperties: true
        repositoryTree:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryTreeItem"
    GitLabFileContent:
      type: object
      required:
        [fileName, filePath, size, content, ref, blobId, commitId, lastCommitId]
      properties:
        fileName:
          type: string
        filePath:
          type: string
        size:
          type: integer
          format: int64
        content:
          type: string
        ref:
          type: string
        blobId:
          type: string
        commitId:
          type: string
        lastCommitId:
          type: string
security:
  - sessionCookie: []